+++ src/main/java/ch/bfh/bti7081/s2015/red/PatientApp/App/Calendar.java
@@ -1,6 +1,8 @@
-package ch.bfh.bti7081.s2015.red.PatientApp.Model;
+package ch.bfh.bti7081.s2015.red.PatientApp.App;
 
 import java.util.ArrayList;
+
+import ch.bfh.bti7081.s2015.red.PatientApp.Model.CalendarEntry;
 
 public class Calendar {
 	private ArrayList<CalendarEntry>entries = new ArrayList<CalendarEntry>();
