+++ src/main/java/ch/bfh/bti7081/s2015/red/PatientApp/App/LifeUp.java
@@ -1,6 +1,8 @@
-package ch.bfh.bti7081.s2015.red.PatientApp.Model;
+package ch.bfh.bti7081.s2015.red.PatientApp.App;
 
 import java.util.ArrayList;
+
+import ch.bfh.bti7081.s2015.red.PatientApp.Model.Activity;
 
 public class LifeUp {
 	ArrayList<Activity> activities = new ArrayList<Activity>();
