+++ src/main/java/ch/bfh/bti7081/s2015/red/PatientApp/App/PatientApp.java
@@ -1,6 +1,11 @@
-package ch.bfh.bti7081.s2015.red.PatientApp.Model;
+package ch.bfh.bti7081.s2015.red.PatientApp.App;
 
 import java.util.ArrayList;
+
+import ch.bfh.bti7081.s2015.red.PatientApp.Model.Activity;
+import ch.bfh.bti7081.s2015.red.PatientApp.Model.CalendarEntry;
+import ch.bfh.bti7081.s2015.red.PatientApp.Model.Emergency;
+import ch.bfh.bti7081.s2015.red.PatientApp.Model.Patient;
 
 public class PatientApp {
 	private Calendar calendar = new Calendar();
