+++ src/main/java/ch/bfh/bti7081/s2015/red/PatientApp/Model/CalendarEntry.java
@@ -2,6 +2,8 @@ package ch.bfh.bti7081.s2015.red.PatientApp.Model;
 
 import java.util.Date;
 
+import ch.bfh.bti7081.s2015.red.PatientApp.Db.Persistable;
+
 import com.google.gson.Gson;
 
 public abstract class CalendarEntry implements Persistable{
