+++ src/main/java/ch/bfh/bti7081/s2015/red/PatientApp/PatientAppStarter.java
@@ -5,7 +5,6 @@ import javax.servlet.annotation.WebServlet;
 import ch.bfh.bti7081.s2015.red.PatientApp.View.NavigatorUI;
 
 import com.vaadin.annotations.VaadinServletConfiguration;
-import com.vaadin.annotations.Widgetset;
 import com.vaadin.server.VaadinServlet;
 
 public class PatientAppStarter {
