+++ src/main/java/ch/bfh/bti7081/s2015/red/PatientApp/View/BaseView.java
@@ -1,8 +1,6 @@
 package ch.bfh.bti7081.s2015.red.PatientApp.View;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-
 import ch.bfh.bti7081.s2015.red.PatientApp.Presenter.ViewListener;
 
 import com.vaadin.ui.VerticalLayout;
